Improve robustness of Level::load_gfx
The previous implementation of Level::load_gfx assumed it would only ever be called on a
newly constructed level object. This would lead potentially to memory leaks if the
function was ever reused. This change allows the function to be self-contained and more
robust. It can now safely handle being called multiple times on the same object by always
clearing the previous background surface before loading a new one.