Bonus cache: Stop asking storage for directions
The Bonus Levels menu was being regenerated on every interaction causing a significant freeze
due to repeated file I/O operations on the main thread. This involved scanning the entire "levels"
directories and parsing info files for every level subset and world map, EVERY SINGLE TIME
the menu was accessed.

This change implements a scan once and scan smart type of strategy to fix the bottleneck. We
technically sacrifice being able to hotswap a level in while the game is running, but it's not
like this was something an ordinary user would be doing anyway on a console or even on the
Dolphin Emulator. This eliminates the freeze and dramatically improves menu responsiveness at
the cost of a negligible increase in initial load time and loss of a feature we never really
had running this via Homebrew.