Made the big cheese less divisive
The mouse cursor's draw function was calculating the width and height of the cursor frame
on EVERY call! Since these dimensions are constant for its lifetime, this was a small but
unnecessary amount of repeated work on the main rendering thread. This eliminates two integer
division operations per frame which should lead to a minor reduction in CPU usage & input latency.