Take the scenic route with Level::draw_bg()
For handling background drawing logic, we now use a centralized method
Level::draw_bg() to handle all scenery rendering. The logic before for
drawing either a background image or a color gradient was duplicated in
world.cpp and gameloop.cpp. This combines the logic into level.cpp and
replaces the redundant code blocks with a clean call to the new function.