Take crashing off the menu
Empty menus no longer serve segfaults. Providing any input prior to a menu without
items would cause a crash. Now we check to ensure the menu is not empty before
attempting to access any of its items, preventing the game from ordering a crash
with a side of null pointer.