The fade that time forgot
For a long time, our special effects department had a fade() function
that had forgotten how to fade. It was designed to create a dramatic
and slow transition into the game's intro, but alas it had a critical
math flaw (integer division) which left it permanently stuck on its
first frame. The lights would never dim and the scene would never
change so it was time to make a change and remove it all.