Tux's lost prologue
Deep in the code's icy data archives, we found the script for the original intro, commented out for over a decade. The
reason it was turned off in the code was likely due to how long it was and how it held players hostage before they
could even start playing. We decided that was going to change today. Also while renovating the theater, we noticed
the stagehands were rebuilding the entire set for the credits every time. The background is now kept handy backstage,
making for a much faster curtain call.