Tux learns to speed read
Tux isn't a patient penguin, especially when browsing for his next
adventure. He was getting frustrated because to see a level's name,
he'd have to read the ENTIRE level file from start to finish. This
was like reading a whole book just to see its title on the cover.

To fix this, we've sent Tux to a speed reading course and now he has
a new skill: get_level_title_fast. Now, just at a glance of a file, we
can find the name and move on.