A clean sweep without the shuffle
Tux noticed that during his most heroic moments, like for instance taking out a whole row
of enemies with a single Mr. Iceblock, the world would sometimes stutter. This happened
because the cleanup crew was inefficient. Whenever an object was destroyed, it made every
other object in its list shuffle forward one by one to fill the empty space. As we don't care
about order, we could simple just have the one at the end move to fill the game instead of moving
each object in the queue. In short, we basically taught the crew a new trick of "swap and pop".