Let's Get Physical
The main action() methods were getting tired of juggling so many responsibilities
at once from checking key presses to calculating bounces. They decided to hire a
specialist named updatePhysics() to handle all the heavy lifting of gravity,
momentum, and colliding with the world. Now the action() methods are free to focus
on their primary jobs which is now simply telling their new helper when it's time to work.