A concrete improvement
The old collision code was a fortress built by a mad architect. The hallways were
an endless echoe of the same design with a central maze that was a treacherous path
which could let Tux at any time phase through the very ground he was standing on. This
renovation tore down the repetitive corridors and replaced them with a single hallway
and rebuilt the central maze with much clearer and more reliable steps. The fortress is
now a safe and sturdy structure where its physics are no longer a work of chaotic magic.